refactor(query): move source expression context into core
